Fix: ensure the moment user uploads the MIDI file that NOTE OFF events are _always_ before NOTE ON event on same tick
Got that file test.mid where I have same note being ON (ON OFF) OFF (same tick). MidiEditor shows such notes as if it was (OFF ON) so I guess this is kind of a standard.

Also a fix for notes changed with form - ensure int values are not strings. Damn js "13" + 20 == "1320"